feat(mealPlanItem): create plan item _ 1

+++ src/main/java/com/hn/nutricarebe/dto/response/NutritionResponse.java
@@ -3,7 +3,6 @@ package com.hn.nutricarebe.dto.response;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
-
 import java.math.BigDecimal;
 
 @Getter
@@ -14,11 +13,11 @@ import java.math.BigDecimal;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class NutritionResponse {
-    Integer kcal;
+    BigDecimal kcal;
     BigDecimal proteinG;
     BigDecimal carbG;
     BigDecimal fatG;
     BigDecimal fiberG;
-    Integer sodiumMg;
+    BigDecimal sodiumMg;
     BigDecimal sugarMg;
 }
